Got objects to spawn infinitely and got pixel accurate hitboxesworking

+++ Food.java
@@ -4,7 +4,7 @@ import java.awt.image.BufferedImage;
 
 public class Food extends Sprite{		//food class for food objects
 	
-	Food(int x, int y, BufferedImage img) {
+	Food(double x, double y, BufferedImage img) {
 		xloc = x;
 		yloc = y;
 		Image = img;
