+++ NestPiece.java
@@ -4,7 +4,7 @@ import java.awt.image.BufferedImage;
 
 public class NestPiece extends Sprite { // nest pieces can be collected
 
-	NestPiece(int x, int y, BufferedImage img) {
+	NestPiece(double x, double y, BufferedImage img) {
 		xloc = x;
 		yloc = y;
 		Image = img;
