+++ Player.java
@@ -4,7 +4,7 @@ import java.awt.image.BufferedImage;
 
 public class Player extends Sprite{
 	
-	Player(int x, int y, BufferedImage img) {
+	Player(double x, double y, BufferedImage img) {
 		xloc = x;
 		yloc = y;
 		Image = img;
